+++ packages/utils/spatial-utils/src/spatial.js
@@ -53,6 +53,40 @@ export function getSourceFromLoader(loader, level) {
   return source;
 }
 
+export const getStatsForResolution = (loader, resolution) => {
+  const { shape, labels } = loader[resolution];
+  const height = shape[labels.indexOf('y')];
+  const width = shape[labels.indexOf('x')];
+  const depth = shape[labels.indexOf('z')];
+  // eslint-disable-next-line no-bitwise
+  const depthDownsampled = Math.max(1, depth >> resolution);
+  // Check memory allocation limits for Float32Array (used in XR3DLayer for rendering)
+  const totalBytes = 4 * height * width * depthDownsampled;
+  return {
+    height, width, depthDownsampled, totalBytes,
+  };
+};
+
+export const canLoadResolution = (loader, resolution) => {
+  const {
+    totalBytes, height, width, depthDownsampled,
+  } = getStatsForResolution(
+    loader,
+    resolution,
+  );
+  const maxHeapSize = window.performance?.memory
+    && window.performance?.memory?.jsHeapSizeLimit / 2;
+  const maxSize = maxHeapSize || (2 ** 31) - 1;
+  // 2048 is a normal texture size limit although some browsers go larger.
+  return (
+    totalBytes < maxSize
+    && height <= 2048
+    && depthDownsampled <= 2048
+    && width <= 2048
+    && depthDownsampled > 1
+  );
+};
+
 /**
  * Helper method to determine whether pixel data is interleaved and rgb or not.
  * @param {object} loader
@@ -76,6 +110,42 @@ export function isRgb(loader, channels) {
 }
 
 // From spatial/utils.js
+
+export function physicalSizeToMatrix(xSize, ySize, zSize, xUnit, yUnit, zUnit) {
+  const hasZPhyscialSize = Boolean(zSize) && Boolean(zUnit);
+  const hasYPhyscialSize = Boolean(ySize) && Boolean(yUnit);
+  const hasXPhyscialSize = Boolean(xSize) && Boolean(xUnit);
+  if (!hasXPhyscialSize || !hasYPhyscialSize) {
+    return (new Matrix4()).identity();
+  }
+  const sizes = [
+    unit(`${xSize} ${xUnit}`.replace('µ', 'u')),
+    unit(`${ySize} ${yUnit}`.replace('µ', 'u')),
+  ];
+  if (hasZPhyscialSize) {
+    sizes.push(unit(`${zSize} ${zUnit}`.replace('µ', 'u')));
+  }
+  // Find the ratio of the sizes to get the scaling factor.
+  const scale = sizes.map(i => divide(i, unit('1 um')));
+
+  // TODO: is this still needed
+  // sizes are special objects with own equals method - see `unit` in declaration
+  if (!sizes[0].equals(sizes[1])) {
+    // Handle scaling in the Y direction for non-square pixels
+    scale[1] = divide(sizes[1], sizes[0]);
+  }
+  // END TODO: is this still needed
+
+  // Add in z dimension needed for Matrix4 scale API.
+  if (!scale[2]) {
+    scale[2] = 1;
+  }
+  // no need to store/use identity scaling
+  if (isEqual(scale, [1, 1, 1])) {
+    return (new Matrix4()).identity();
+  }
+  return new Matrix4().scale([...scale]);
+}
 
 function getMetaWithTransformMatrices(imageMeta, imageLoaders) {
   // Do not fill in transformation matrices if any of the layers specify one.
@@ -367,11 +437,11 @@ export function getNgffAxesForTiff(dimOrder) {
 /**
  * Convert an array of coordinateTransformations objects to a 16-element
  * plain JS array using Matrix4 linear algebra transformation functions.
- * @param {object[]} coordinateTransformations List of objects matching the
+ * @param {object[]|undefined} coordinateTransformations List of objects matching the
  * OME-NGFF v0.4 coordinateTransformations spec.
- * @param {object[]} axes Axes in OME-NGFF v0.4 format, objects
+ * @param {object[]|undefined} axes Axes in OME-NGFF v0.4 format, objects
  * with { type, name }.
- * @returns {Array<number>} Array of 16 numbers representing the Matrix4.
+ * @returns {Matrix4} Array of 16 numbers representing the Matrix4.
  */
 export function coordinateTransformationsToMatrix(coordinateTransformations, axes) {
   let mat = (new Matrix4()).identity();
@@ -415,6 +485,15 @@ export function coordinateTransformationsToMatrix(coordinateTransformations, axe
     });
   }
   return mat;
+}
+
+export function hexToRgb(hex) {
+  const result = /^#?([A-F\d]{2})([A-F\d]{2})([A-F\d]{2})$/i.exec(hex);
+  return [
+    parseInt(result[1].toLowerCase(), 16),
+    parseInt(result[2].toLowerCase(), 16),
+    parseInt(result[3].toLowerCase(), 16),
+  ];
 }
 
 /**