+++ packages/view-types/layer-controller/src/LayerController.js
@@ -1,7 +1,12 @@
 import React, { useState, useRef, useEffect, useId } from 'react';
 import { viv } from '@vitessce/gl';
 import {
-  GLOBAL_LABELS, getSourceFromLoader, isRgb,
+  GLOBAL_LABELS,
+  getSourceFromLoader,
+  isRgb,
+  getMultiSelectionStats,
+  DOMAINS,
+  canLoadResolution,
 } from '@vitessce/spatial-utils';
 
 import {
@@ -15,8 +20,6 @@ import {
   VisibilityOff as VisibilityOffIcon,
 } from '@material-ui/icons';
 
-
-import { canLoadResolution } from '@vitessce/utils';
 import LayerOptions from './LayerOptions.js';
 import VolumeOptions from './VolumeOptions.js';
 import {
@@ -25,9 +28,7 @@ import {
   useOverflowEllipsisGridStyles,
   useAccordionStyles,
 } from './styles.js';
-import { getMultiSelectionStats } from './utils.js';
 
-import { DOMAINS } from './constants.js';
 
 function TabPanel(props) {
   const {