+++ packages/vit-s/src/shared-mui/components.js
@@ -8,6 +8,7 @@ import {
   ClickAwayListener,
   Fade,
 } from '@material-ui/core';
+import clsx from 'clsx';
 import { useVitessceContainer } from '../hooks.js';
 
 const useStyles = makeStyles(() => ({
@@ -30,6 +31,8 @@ export function PopperMenu(props) {
     children,
     buttonClassName,
     placement = 'bottom-end',
+    withPaper = true,
+    containerClassName,
     'aria-label': ariaLabel,
   } = props;
   const classes = useStyles();
@@ -49,7 +52,7 @@ export function PopperMenu(props) {
   const getTooltipContainer = useVitessceContainer(anchorRef);
 
   return (
-    <div ref={anchorRef} className={classes.container}>
+    <div ref={anchorRef} className={clsx(classes.container, containerClassName)}>
       <IconButton
         aria-describedby={id}
         onClick={handleClick}
@@ -72,9 +75,11 @@ export function PopperMenu(props) {
         {({ TransitionProps }) => (
           <ClickAwayListener onClickAway={handleClose}>
             <Fade {...TransitionProps} timeout={100}>
-              <Paper elevation={4} className={classes.paper}>
-                <MenuList>{children}</MenuList>
-              </Paper>
+              {withPaper ? (
+                <Paper elevation={4} className={classes.paper}>
+                  <MenuList>{children}</MenuList>
+                </Paper>
+              ) : children}
             </Fade>
           </ClickAwayListener>
         )}